RAG/LangChain_RAG/requirements.txt
streamlit
langchain
langchain_community
langchain_core
langgraph
chromadb
typing-extensions
bs4
PyPDF2
python-docx
sentence-transformers
